Moved content out of ARE, once I'm done cleaning up the entity class tree, I'll work on simplifying the active and passive systems, then redo the backend. Definitely want to try and get in a couple days of power-programming a Java version of GEARS before classes start again, after I finally am no longer sick.
