+++ src/main/java/entities/AbstractPlayer.java
@@ -13,6 +13,7 @@ import controllers.World;
 import customizables.characterClass.CharacterStatName;
 import static java.lang.System.out;
 import java.util.Arrays;
+import util.Direction;
 import util.SafeList;
 
 /**
@@ -24,13 +25,23 @@ import util.SafeList;
 public abstract class AbstractPlayer extends AbstractReactiveEntity{
 	private final String name;
 	private Color color;
-	
-	private final DamageBacklog log;
-	
-	private final MeleeActive slash;
     
+    /*
+	 * (focusX, focusY) is a point that the entity is trying to reach
+	 */
+	private int focusX;
+	private int focusY;
+	private boolean hasFocus;
+    
+    private Direction knockbackDir;
+    private int knockbackMag;
+    private int knockbackDur;
+    
+    private int lastHitById; //the useId of the last projectile that hit this player
+	
+    private final MeleeActive slash;
+	private final DamageBacklog log;
 	private final SafeList<AbstractStatus> statuses; //change to hashtable
-	private int lastHitById; //the useId of the last projectile that hit this player
     //both players and AI need to find paths, given the current controls
     private Path path;
     
@@ -43,13 +54,21 @@ public abstract class AbstractPlayer extends AbstractReactiveEntity{
 		name = n;
         color = Color.black;
         
+        focusX = 0;
+        focusY = 0;
+        hasFocus = false;
+        
+        knockbackDir = null;
+        knockbackMag = 0;
+        knockbackDur = 0;
+        
         slash = (MeleeActive)AbstractActive.getActiveByName("Slash");
 		slash.setUser(this);
         log = new DamageBacklog(this, minLifeSpan);
-		
-		setRadius(RADIUS);
         path = null;
         statuses = new SafeList<>();
+        
+        setRadius(RADIUS);
 	}
 	
 	public final String getName(){
@@ -61,6 +80,25 @@ public abstract class AbstractPlayer extends AbstractReactiveEntity{
     
 	public DamageBacklog getLog(){
 		return log;
+	}
+    
+    //focus related methods
+	public final void setFocus(int xCoord, int yCoord){
+		focusX = xCoord;
+		focusY = yCoord;
+		hasFocus = true;
+	}
+	public final void setFocus(AbstractEntity e){
+		setFocus(e.getX(), e.getY());
+	}
+	public final void turnToFocus(){
+		turnTo(focusX, focusY);
+	}
+	public boolean withinFocus(){
+		// returns if has reached focal point
+		boolean withinX = Math.abs(getX() - focusX) < getSpeed();
+		boolean withinY = Math.abs(getY() - focusY) < getSpeed();
+		return withinX && withinY;
 	}
     
     public void setPath(int x, int y){
@@ -76,6 +114,17 @@ public abstract class AbstractPlayer extends AbstractReactiveEntity{
     }
     public Path getPath(){
         return path;
+    }
+    
+    public final void knockBack(int mag, Direction d, int dur){
+        /**
+         * @param mag : the total distance this entity will be knocked back
+         * @param d : the direction this entity is knocked back
+         * @param dur : the number of frames this will be knocked back for
+         */
+        knockbackMag = mag / dur;
+        knockbackDir = d;
+        knockbackDur = dur;
     }
     
 	public void inflict(AbstractStatus newStat){
@@ -134,8 +183,38 @@ public abstract class AbstractPlayer extends AbstractReactiveEntity{
         path = null;
 		lastHitById = -1;
         
+        hasFocus = false;
+        knockbackDir = null;
+        knockbackMag = 0;
+        knockbackDur = 0;
+        
         playerInit();
 	}
+    
+    @Override
+    public void updateMovement(){
+        if(hasFocus){
+			if(withinFocus()){
+				hasFocus = false;
+				setMoving(false);
+			}else{
+				turnToFocus();
+				setMoving(true);
+			}
+		}
+        if(knockbackDir != null){
+            //cannot move if being knocked back
+            setX((int) (getX() + knockbackMag * knockbackDir.getXMod()));
+            setY((int) (getY() + knockbackMag * knockbackDir.getYMod()));
+            knockbackDur--;
+            if(knockbackDur == 0){
+                knockbackDir = null;
+            }
+        } else {
+            super.updateMovement();
+        }   
+        clearSpeedFilter();
+    }
 	
     @Override
 	public void update(){
